# db: enable TrySeekUsingNext after Next in external iters